add alt on author avatar
